--- backend/api/run.py
@@ -1,5 +1,0 @@
-from app import app
-from app import routes
-
-if __name__ == "__main__":
-    app.run(port=5000, debug=True)
