+++ backend/models/course.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from datetime import datetime
 
-from models import ClassLevel, Credit
+from backend.models import ClassLevel, Credit
 
 
 @dataclass(frozen=True)
