+++ backend/models/course_department.py
