+++ backend/models/professor_course_rating.py
@@ -1,10 +1,14 @@
 from dataclasses import dataclass
 from datetime import datetime
 
+from backend.models import Rate
+
 
 @dataclass(frozen=True)
-class Ucore:
+class ProfessorCourseRating:
     id: int
-    name: str
+    professor_id: int
+    course_id: int
+    rating: Rate
     created_at: datetime
     updated_at: datetime
