+++ backend/models/rating.py
@@ -1,14 +1,16 @@
 from dataclasses import dataclass
 from datetime import datetime
 
-from schema.models import Rate
+from backend.models import Rate
 
 
 @dataclass(frozen=True)
-class ProfessorCourseRating:
+class Rating:
     id: int
     professor_id: int
     course_id: int
-    rating: Rate
+    rmp_quality: Rate
+    rmp_difficulty: Rate
+    rmp_comment: str
     created_at: datetime
     updated_at: datetime
