--- backend/scrape/run.py
@@ -1,2 +1,0 @@
-if __name__ == "__main__":
-    pass  # TODO: Set up the scraping process
