--- backend/scrape/test/test.py
@@ -1,6 +1,0 @@
-from scrape.web_scraper import scrape_course_info
-import pytest
-
-
-def test_scrape_course_info():
-    raise NotImplementedError("Test not implemented")
